TASK: Split Flow Log to separate package
Creates the new `neos/flow-log` package containing the original
log infrastructure with the original namespace (Neos\Flow\Log). So
nothing should change for userland code.

The choosen way to split allows for transitioning to a new namespace
while keeping the old classes.

The `SystemLoggerInterface` and `SecurityLoggerInterface` are kept
in Flow as they have not much meaning in the Logger package.